requirements.txt
Flask==2.3.3
Pillow==9.5.0
opencv-python-headless==4.8.0.76
numpy==1.24.0
google-cloud-secret-manager==2.16.2
google-auth>=2.15.0
google-auth-httplib2>=0.1.0
google-api-python-client==2.93.0
gunicorn==20.1.0
